requirements.txt
openai
streamlit
streamlit-extras
streamlit-chat
streamlit-lottie
Pillow
langchain